gluster-nagios-monitoring: Fix undefined cluster issue
This patch fixes the issue of undefined cluster being
passed to ClusterService#getClusters which results in
plugin being removed from service and hence no graphs
appear.
This change was neccessary due to a recent refactoring
patch that refactored the service method to take a
cluster object and although the angular treats passing
nothing to a function expecting something as equivalent
to passing undefined object, the undefined check in the
service is capable of handling a undefines string and
replacing that with angular#isUndefined solves our
purpose but that needs heavy refactoring all across
the plugin code and hence this patch takes the approach
of sending a undefined string.

Change-Id: Ib22fdac17c8ec8bb58902f4eb02231d815437cb1
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1228291
Signed-off-by: Anmol Babu <anbabu@redhat.com>